extract team dropdown into components
