+++ src/core/dashboard/teams/team-menu/DeleteTeam.tsx
@@ -1,7 +1,6 @@
-import { type $Enums } from '@prisma/client';
-import { MenuIcon, PencilIcon, TrashIcon } from 'lucide-react';
 import { useState } from 'react';
-import { Button } from '~/components/ui/Button';
+import type { Role, Team } from './TeamDropdown';
+import { api } from '~/utils/api';
 import {
   Dialog,
   DialogContent,
@@ -10,74 +9,12 @@ import {
   DialogTitle,
   DialogTrigger,
 } from '~/components/ui/Dialog';
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuGroup,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuTrigger,
-} from '~/components/ui/DropdownMenu';
+import { DropdownMenuItem } from '~/components/ui/DropdownMenu';
+import { TrashIcon } from 'lucide-react';
 import { Input } from '~/components/ui/Input';
-import { cn } from '~/lib/utils';
-import { type RouterOutputs, api } from '~/utils/api';
-import { useNewTeam } from './new-team/newTeamStore';
+import { Button } from '~/components/ui/Button';
 
-type Team = RouterOutputs['team']['memberOfRoles'][number];
-type Role = $Enums.RosterRole | Exclude<$Enums.TeamRole, 'member'>;
-
-export default function TeamDropdown({
-  team,
-  roles,
-  className,
-}: {
-  team: Team;
-  roles: Role[];
-  className?: string;
-}) {
-  const [menuOpened, setMenuOpened] = useState(false);
-  const { setSheetOpened, setEditMode, setEditModeTeamId, setData } =
-    useNewTeam();
-
-  const handleEditTeam = () => {
-    setSheetOpened(true);
-    setEditMode(true);
-    setEditModeTeamId(team.id);
-
-    setData({
-      name: team.name,
-      abbreviation: team.abbreviation,
-      password: '',
-      profilePicture: null,
-    });
-  };
-
-  return (
-    <DropdownMenu open={menuOpened} onOpenChange={setMenuOpened} modal={false}>
-      <DropdownMenuTrigger asChild>
-        <Button
-          variant='ghost'
-          className={cn('h-8 w-8 px-0', className)}
-          aria-label={`${menuOpened ? 'close' : 'open'} the menu`}
-        >
-          <MenuIcon className='h-5 w-5' />
-        </Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align='end' className='w-56'>
-        <DropdownMenuLabel>Manage team</DropdownMenuLabel>
-        <DropdownMenuGroup>
-          <DropdownMenuItem onClick={handleEditTeam}>
-            <PencilIcon className='mr-2 h-4 w-4' />
-            <span>Edit team</span>
-          </DropdownMenuItem>
-          <DeleteTeamDialog team={team} roles={roles} />
-        </DropdownMenuGroup>
-      </DropdownMenuContent>
-    </DropdownMenu>
-  );
-}
-
-function DeleteTeamDialog({
+export default function DeleteTeamDialog({
   team: { name, id: teamId },
   roles,
 }: {
